Refactor folders into clean architecture
