+++ app/src/main/java/com/example/fitnesslog/exercise/data/dao/SetDao.kt
@@ -1,11 +1,11 @@
-package com.example.fitnesslog.set.data.dao
+package com.example.fitnesslog.exercise.data.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Transaction
 import androidx.room.Update
-import com.example.fitnesslog.set.data.model.SetTemplate
+import com.example.fitnesslog.exercise.domain.model.SetTemplate
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.firstOrNull
