+++ app/src/main/java/com/example/fitnesslog/exercise/domain/model/SetTemplate.kt
@@ -1,4 +1,4 @@
-package com.example.fitnesslog.set.data.model
+package com.example.fitnesslog.exercise.domain.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
@@ -6,7 +6,6 @@ import androidx.room.ForeignKey
 import androidx.room.Ignore
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.example.fitnesslog.exercise.data.model.ExerciseTemplate
 
 
 /**
