+++ app/src/main/java/com/example/fitnesslog/exercise/domain/model/WorkoutSessionExercise.kt
@@ -1,11 +1,11 @@
-package com.example.fitnesslog.exercise.data.model
+package com.example.fitnesslog.exercise.domain.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.example.fitnesslog.workout.data.model.WorkoutSession
+import com.example.fitnesslog.workout.domain.model.WorkoutSession
 
 /**
  * Represents an exercise within a specific workout session.
