+++ app/src/main/java/com/example/fitnesslog/exercise/domain/model/WorkoutTemplateExercise.kt
@@ -1,11 +1,11 @@
-package com.example.fitnesslog.exercise.data.model
+package com.example.fitnesslog.exercise.domain.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.example.fitnesslog.workout.data.model.WorkoutTemplate
+import com.example.fitnesslog.workout.domain.model.WorkoutTemplate
 
 /**
  * Represents the association between a workout template and its exercises.
