+++ app/src/main/java/com/example/fitnesslog/workout/domain/model/WorkoutSession.kt
@@ -1,4 +1,4 @@
-package com.example.fitnesslog.workout.data.model
+package com.example.fitnesslog.workout.domain.model
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
