+++ app/src/main/java/com/example/fitnesslog/workout/view/fragment/WorkoutHomeFragment.kt
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [WorkoutFragment.newInstance] factory method to
+ * Use the [WorkoutHomeFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class WorkoutFragment : Fragment() {
+class WorkoutHomeFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class WorkoutFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_workout, container, false)
+        return inflater.inflate(R.layout.fragment_workout_home, container, false)
     }
 
     companion object {
@@ -50,7 +50,7 @@ class WorkoutFragment : Fragment() {
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            WorkoutFragment().apply {
+            WorkoutHomeFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
