+++ app/src/main/res/layout/fragment_workout_home.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".workout.view.fragment.WorkoutFragment">
+    tools:context=".workout.view.fragment.WorkoutHomeFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
