+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -4,16 +4,16 @@
     <item
         android:id="@+id/historyFragment"
         android:title="@string/label_history"
-        android:icon="@drawable/baseline_history_24"/>
+        android:icon="@drawable/baseline_history_24" />
 
     <item
-        android:id="@+id/workoutFragment"
+        android:id="@+id/workoutHomeFragment"
         android:title="@string/label_workout"
-        android:icon="@drawable/outline_fitness_center_24"/>
+        android:icon="@drawable/outline_fitness_center_24" />
 
     <item
         android:id="@+id/programFragment"
         android:title="@string/label_program"
-        android:icon="@drawable/baseline_folder_open_24"/>
+        android:icon="@drawable/baseline_folder_open_24" />
 
 </menu>
